+++ src/QueryBuilder/QueryBuilder/helpers/query.js
@@ -326,8 +326,12 @@ export const getSourceValue = ({
   getParamsSource,
   getDataOptions,
 }) => {
-  // if initial value provided, fill the source with it
-  if (initialValues) {
+  // if initial value provided, and it has some items, fill the source with it
+  const hasInitialValues = Object
+    .values(initialValues ?? {})
+    .some(val => (Array.isArray(val) ? val.length > 0 : Object.keys(val).length > 0));
+
+  if (hasInitialValues) {
     return fqlQueryToSource({
       initialValues,
       fieldOptions,